packages/template-validation/pyproject.toml
[project]
name = "template-validation"
version = "0.1.0"
description = "Validation tools for gptme-agent-template forks"
authors = [
    {name = "Lofty", email = "lofty@loftybuilder.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
template-validation = "template_validation.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=template_validation --cov-report=term-missing"
